+++ build_runner/lib/src/build_script_generate/build_script_generate.dart
@@ -4,7 +4,7 @@
 
 import 'dart:async';
 
-import 'package:build/build.dart';
+import 'package:build/build.dart' show BuilderOptions;
 import 'package:build_config/build_config.dart';
 import 'package:build_runner_core/build_runner_core.dart';
 import 'package:code_builder/code_builder.dart';
@@ -19,8 +19,10 @@ import 'builder_ordering.dart';
 const scriptLocation = '$entryPointDir/build.dart';
 const scriptSnapshotLocation = '$scriptLocation.snapshot';
 
-Future<String> generateBuildScript() => logTimedAsync(
-    Logger('Entrypoint'), 'Generating build script', _generateBuildScript);
+final _log = Logger('Entrypoint');
+
+Future<String> generateBuildScript() =>
+    logTimedAsync(_log, 'Generating build script', _generateBuildScript);
 
 Future<String> _generateBuildScript() async {
   final builders = await _findBuilderApplications();
@@ -34,7 +36,13 @@ Future<String> _generateBuildScript() async {
         _main()
       ]));
   final emitter = DartEmitter(Allocator.simplePrefixing());
-  return DartFormatter().format('${library.accept(emitter)}');
+  try {
+    return DartFormatter().format('${library.accept(emitter)}');
+  } on FormatterException {
+    _log.severe('Generated build script could not be parsed.\n'
+        'This is likely caused by a misconfigured builder definition.');
+    throw CannotBuildException();
+  }
 }
 
 /// Finds expressions to create all the `BuilderApplication` instances that
@@ -207,7 +215,7 @@ String _buildScriptImport(String import) {
   if (import.startsWith('package:')) {
     return import;
   } else if (import.startsWith('../') || import.startsWith('/')) {
-    log.warning('The `../` import syntax in build.yaml is now deprecated, '
+    _log.warning('The `../` import syntax in build.yaml is now deprecated, '
         'instead do a normal relative import as if it was from the root of '
         'the package. Found `$import` in your `build.yaml` file.');
     return import;